Assorted functions from the remaining nonmatchings (#196)
* scr_trans

* hex to uppercase

* Bg_Family_Set

* Bg_Family_Set_2

* Bg_Family_Set_2_appoint

* chase_start_check

* ake_Family_Set2

* Bg_Family_Set_appoint

* cal_move_quantity3

* cal_all_speed_data

* cal_initial_speed

* cal_initial_speed_y, cal_delta_speed

* cal_top_of_position_y

* cal_move_dir_forecast

* remake_2_10

* grade_check_work_1st_init

* get_target_att_position

* set_paring_status

* eff_hit_check_sub

* added #if defined(TARGET_PS2) declares for grade_check_work_1st_init to
Entry_01_Sub and Sel_PL_3rd

* sync_fam_set3

* seqsGetSprMax - lol

* effect_at_vs_effect_dm

* setup_dm_rl_pldm

* these should be WORK_Other because it's effect vs effect

* player_at_vs_effect_dm

* bg_etc_write

* zoom_ud_check

* player_at_vs_player_dm

* missed some stubs and includes

* pr suggestions

+++ include/sf33rd/Source/Game/Grade.h
@@ -9,7 +9,7 @@ extern JudgeGals judge_gals[2];
 extern GradeFinalData judge_final[2][2]; // size: 0x2B0, address: 0x5E3070
 extern GradeData judge_item[2][2];       // size: 0x130, address: 0x5E3320
 
-void grade_check_work_1st_init(s32 ix, s32 ix2);
+void grade_check_work_1st_init(s16 ix, s16 ix2);
 void grade_check_work_stage_init(s16 ix);
 void grade_check_work_round_init(s16 ix);
 void grade_makeup_final_parameter(s16 ix, s16 pt);
